requirements.txt
beautifulsoup4==4.8.0
click==7.0
lxml==4.4.1
nltk==3.4.5
numpy==1.17.0
opencv-python==4.1.0.25
pdfminer.six==20181108
Pillow==6.1.0
PyPDF2==1.26.0
scipy==1.3.1
spacy==2.1.8